step1-7-refactor-for-reducer
